Support up to 64 AFC lanes with per-filament T# remap
Scale OrcaSlicer's AMS/filament plumbing to handle Klipper/AFC setups that
use up to 64 lanes with sparse T<N> tool-number assignments.

- New per-filament config filament_map_tool_number carries the AFC T<N>
  override from the printer's filament_ams_list through project_config.
- GCodeOutputStream::write now applies an optional transform that rewrites
  emitted T<filament_id> lines to T<afc_tool_number>. The transform is only
  applied to bytes written to disk; the GCodeProcessor analyser still sees
  filament indices so the in-app preview and per-filament estimates remain
  correct.
- MoonrakerPrinterAgent groups lanes by AFC unit for the UI (box0, box1,
  ...) and emits each tray's AFC T<N> value in afc_tool_number so the UI
  layout is decoupled from G-code tool routing.
- Widened ams_exist_bits / tray_exist_bits (and the StatusPanel trackers
  that parse them) from long to uint64_t so the 64-lane payload fits.
- Tool-number nibble in the info hex word is no longer used for AFC
  routing; extruder_tool_number is carried explicitly per tray.

+++ src/slic3r/GUI/StatusPanel.cpp
@@ -3412,13 +3412,13 @@ void StatusPanel::update_ams(MachineObject *obj)
                 int         tray_id_int = atoi(tray_id.c_str());
                 // new protocol
                 if (ams_id_int < 128) {
-                    if ((obj->tray_reading_bits & (1 << (ams_id_int * 4 + tray_id_int))) != 0) {
+                    if ((obj->tray_reading_bits & (1ULL << (ams_id_int * 4 + tray_id_int))) != 0) {
                         m_ams_control->PlayRridLoading(ams_id, tray_id);
                     } else {
                         m_ams_control->StopRridLoading(ams_id, tray_id);
                     }
                 } else {
-                    int check_flag = (1 << (16 + ams_id_int - 128));
+                    uint64_t check_flag = (1ULL << (16 + ams_id_int - 128));
                     if ((obj->tray_reading_bits & check_flag) != 0) {
                         m_ams_control->PlayRridLoading(ams_id, tray_id);
                     } else {
